requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
fastapi==0.115.12
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
instagrapi==2.1.5
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
MarkupSafe==3.0.2
narwhals==1.44.0
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytrends==4.9.2
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@614d4c2029a62d348ca56598f87c425966aaec66
sodapy==2.2.0
soupsieve==2.7
starlette==0.46.2
streamlit==1.46.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.1
